+++ sulong/mx.sulong/mx_testsuites.py
@@ -30,6 +30,7 @@
 from __future__ import print_function
 
 import fnmatch
+import glob
 
 import mx
 import mx_unittest
@@ -262,4 +263,30 @@ class ExternalTestSuite(SulongTestSuite):  # pylint: disable=too-many-ancestors
         roots = [d.get_path(resolve=True) for d in self.buildDependencies if d.isPackedResourceLibrary()]
         env['VPATH'] = ':'.join(roots)
         env['TESTFILE'] = self.getTestFile()
+        return env
+
+
+class GlobNativeProject(mx.NativeProject):
+    def __init__(self, suite, name, deps, workingSets, subDir, results=None, output=None, **args):
+        projectDir = args.pop('dir', None)
+        if projectDir:
+            d = os.path.join(suite.dir, projectDir)
+        elif subDir is None:
+            d = os.path.join(suite.dir, name)
+        else:
+            d = os.path.join(suite.dir, subDir, name)
+        super(GlobNativeProject, self).__init__(suite, name, subDir, [], deps, workingSets, results, output, d, **args)
+
+    def getResults(self, replaceVar=mx_subst.results_substitutions):
+        results = super(GlobNativeProject, self).getResults(replaceVar)
+
+        def _glob(pathname):
+            return glob.glob(pathname) or mx.abort(
+                'Glob pattern "{}" did not return any file'.format(os.path.relpath(pathname, self.getOutput(replaceVar))))
+        results = [expanded for entry in results for expanded in _glob(entry)]
+        return results
+
+    def getBuildEnv(self, replaceVar=mx_subst.path_substitutions):
+        env = super(GlobNativeProject, self).getBuildEnv(replaceVar=replaceVar)
+        env['SRC_DIR'] = self.dir
         return env