+++ sulong/projects/com.oracle.truffle.llvm.parser/src/com/oracle/truffle/llvm/parser/macho/MachORPathCommand.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2019, Oracle and/or its affiliates.
+ * Copyright (c) 2017, 2019, Oracle and/or its affiliates.
  *
  * All rights reserved.
  *
@@ -27,26 +27,37 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.oracle.truffle.llvm.nodes.intrinsics.interop;
+package com.oracle.truffle.llvm.parser.macho;
 
-import com.oracle.truffle.api.dsl.Cached;
-import com.oracle.truffle.api.dsl.CachedContext;
-import com.oracle.truffle.api.dsl.NodeChild;
-import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.llvm.nodes.intrinsics.llvm.LLVMIntrinsic;
-import com.oracle.truffle.llvm.runtime.LLVMContext;
-import com.oracle.truffle.llvm.runtime.LLVMLanguage;
-import com.oracle.truffle.llvm.runtime.nodes.api.LLVMExpressionNode;
+public final class MachORPathCommand extends MachOLoadCommand {
 
-@NodeChild(type = LLVMExpressionNode.class)
-public abstract class LLVMLoadLibrary extends LLVMIntrinsic {
+    private final String name;
 
-    @Specialization
-    protected Object doIntrinsic(Object value,
-                    @CachedContext(LLVMLanguage.class) LLVMContext context,
-                    @Cached("createReadString()") LLVMReadStringNode readId) {
-        String name = readId.executeWithTarget(value);
-        context.addExternalLibrary(name, true, "<truffle_load_library>");
-        return null;
+    private MachORPathCommand(int cmd, int cmdSize, String name) {
+        super(cmd, cmdSize);
+        this.name = name;
     }
+
+    public String toString() {
+        return getClass().getSimpleName() + "<" + name + ">";
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public static MachORPathCommand create(MachOReader buffer) {
+        int pos = buffer.getPosition();
+        int cmd = buffer.getInt();
+        assert cmd == MachOLoadCommand.LC_RPATH;
+        int cmdSize = buffer.getInt();
+        int offset = buffer.getInt();
+
+        buffer.setPosition(pos + offset);
+        String path = getString(buffer, cmdSize - offset);
+        buffer.setPosition(pos + cmdSize);
+
+        return new MachORPathCommand(cmd, cmdSize, path);
+    }
+
 }