+++ sulong/projects/com.oracle.truffle.llvm.parser/src/com/oracle/truffle/llvm/parser/model/ModelModule.java
@@ -30,6 +30,7 @@
 package com.oracle.truffle.llvm.parser.model;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -58,6 +59,8 @@ public final class ModelModule {
     private final ArrayList<FunctionDeclaration> declares = new ArrayList<>();
     private final ArrayList<FunctionDefinition> defines = new ArrayList<>();
     private final ArrayList<TargetInformation> targetInfo = new ArrayList<>();
+    private final ArrayList<String> libraries = new ArrayList<>();
+    private final ArrayList<String> paths = new ArrayList<>();
     private final HashMap<LLVMSourceSymbol, SymbolImpl> sourceGlobals = new HashMap<>();
     private final HashMap<LLVMSourceStaticMemberType, SymbolImpl> sourceStaticMembers = new HashMap<>();
     private final HashMap<FunctionDefinition, LazyFunctionParser> lazyFunctionParsers = new HashMap<>();