+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/LLVMContext.java
@@ -29,6 +29,7 @@
  */
 package com.oracle.truffle.llvm.runtime;
 
+import java.io.PrintStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.SecureRandom;
@@ -43,6 +44,7 @@ import java.util.Objects;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
+import com.oracle.truffle.api.CompilerDirectives;
 import org.graalvm.collections.EconomicMap;
 
 import com.oracle.truffle.api.CallTarget;
@@ -447,19 +449,27 @@ public final class LLVMContext {
         return Paths.get(lib);
     }
 
+    public ExternalLibrary addExternalLibrary(String lib, boolean isNative, Object reason) {
+        return addExternalLibrary(lib, isNative, reason, DefaultLibraryLocator.INSTANCE);
+    }
+
     /**
      * @return null if already loaded
      */
-    public ExternalLibrary addExternalLibrary(String lib, boolean isNative) {
+    public ExternalLibrary addExternalLibrary(String lib, boolean isNative, Object reason, LibraryLocator locator) {
         CompilerAsserts.neverPartOfCompilation();
         if (isInternalLibrary(lib)) {
             // Disallow loading internal libraries explicitly.
             return null;
         }
-        Path path = locateExternalLibrary(lib);
+        Path path = locator.locate(this, lib, reason);
         ExternalLibrary newLib = ExternalLibrary.external(path, isNative);
         ExternalLibrary existingLib = addExternalLibrary(newLib);
-        return existingLib == newLib ? newLib : null;
+        if (existingLib == newLib) {
+            return newLib;
+        }
+        traceLoader("library already located: %s\n", existingLib.path);
+        return null;
     }
 
     private boolean isInternalLibrary(String lib) {
@@ -510,10 +520,18 @@ public final class LLVMContext {
         // failures at the moment, because the library path is not always set correctly
     }
 
+    List<Path> getLibraryPaths() {
+        // TODO (je) should this be unmodifiable?
+        return libraryPaths;
+    }
+
     @TruffleBoundary
-    private Path locateExternalLibrary(String lib) {
+    private Path locateExternalLibrary(String lib, Object reason, List<String> localPaths) {
+        traceLoader("\n");
+        traceLoaderFind(lib, reason);
         Path libPath = Paths.get(lib);
         if (libPath.isAbsolute()) {
+            traceLoaderTry(libPath);
             if (libPath.toFile().exists()) {
                 return libPath;
             } else {
@@ -521,13 +539,27 @@ public final class LLVMContext {
             }
         }
 
+        // search global paths
+        traceLoaderSearchPath(libraryPaths);
         for (Path p : libraryPaths) {
             Path absPath = Paths.get(p.toString(), lib);
+            traceLoaderTry(absPath);
             if (absPath.toFile().exists()) {
                 return absPath;
             }
         }
-
+        if (localPaths != null) {
+            // search file local paths
+            traceLoaderSearchPath(localPaths, reason);
+            for (String p : localPaths) {
+                Path absPath = Paths.get(p, lib);
+                traceLoaderTry(absPath);
+                if (absPath.toFile().exists()) {
+                    return absPath;
+                }
+            }
+        }
+        traceLoaderTry(libPath);
         return libPath;
     }
 
@@ -895,4 +927,50 @@ public final class LLVMContext {
             return scopes.contains(scope);
         }
     }
+
+    @CompilationFinal private boolean traceLoaderEnabled;
+    @CompilationFinal private PrintStream traceLoaderStream;
+
+    private void cacheTrace() {
+        if (traceLoaderStream == null) {
+            CompilerDirectives.transferToInterpreterAndInvalidate();
+            traceLoaderStream = SulongEngineOption.getStream(getEnv().getOptions().get(SulongEngineOption.LD_DEBUG));
+            traceLoaderEnabled = SulongEngineOption.isTrue(getEnv().getOptions().get(SulongEngineOption.LD_DEBUG));
+        }
+    }
+
+    private boolean ldDebugEnabled() {
+        cacheTrace();
+        return traceLoaderEnabled;
+    }
+
+    private PrintStream ldDebugStream() {
+        cacheTrace();
+        return traceLoaderStream;
+    }
+
+    public void traceLoaderFind(Object lib, Object reason) {
+        traceLoader("find external library=%s; needed by %s\n", lib, reason);
+    }
+
+    public void traceLoaderTry(Object file) {
+        traceLoader("  trying file=%s\n", file);
+    }
+
+    public void traceLoaderSearchPath(List<?> paths) {
+        traceLoader(" search path=%s\n", paths);
+    }
+
+    public void traceLoaderSearchPath(List<?> paths, Object reason) {
+        traceLoader(" search path=%s (local path from %s)\n", paths, reason);
+    }
+
+    @TruffleBoundary
+    public void traceLoader(String format, Object... args) {
+        if (ldDebugEnabled()) {
+            ldDebugStream().printf("lli(%x): ", System.identityHashCode(this));
+            ldDebugStream().printf(format, args);
+        }
+    }
+
 }