+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/linker/A.c
@@ -1,0 +1,9 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int globalA = 1;
+
+int methodA(int a, int b) {
+  printf("A\n");
+  return a + b + globalA;
+}