+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/linker/B.c
@@ -1,0 +1,9 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int globalB = 2;
+
+int methodB(int a, int b) {
+  printf("B\n");
+  return a - b - globalB;
+}