+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/linker/E.c
@@ -1,0 +1,9 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int globalE = 5;
+
+int methodE(int a, int b) {
+  printf("E\n");
+  return (a << b) + globalE;
+}