+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/linker/FileLocalA.c
@@ -1,0 +1,9 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+static int globalA = 8;
+
+static int methodA(int a, int b) {
+  printf("FileLocalA\n");
+  return a * b + globalA;
+}