+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/rpath/BarB.c
@@ -1,0 +1,7 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int bar(int a, int b) {
+  printf("barB\n");
+  return a + b + 37;
+}