+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/rpath/BarC.c
@@ -1,0 +1,7 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int bar(int a, int b) {
+  printf("barC\n");
+  return a - b - 42;
+}