+++ sulong/tests/com.oracle.truffle.llvm.tests.linker/rpath/FooA.c
@@ -1,0 +1,9 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+int bar(int a, int b);
+
+int foo(int a, int b) {
+  printf("fooA\n");
+  return (b - a) + bar(3, 9);
+}